% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateImageData.R
\name{CreateImageData}
\alias{CreateImageData}
\title{U-Method CreateImageData function}
\usage{
CreateImageData(bc_matrix, poaraq, markers, project = NULL, print.table = F)
}
\arguments{
\item{bc_matrix}{10x filtered matrix, can be loaded by Read10X(".../Visium_HD__binned_outputsXum/filtered_feature_bc_matrix/"), X in (2,8,16).}

\item{poaraq}{poaraq file can be loaded by read_parquet(".../Visium_HD__binned_outputX2um/spatial/tissue_positions.parquet")}

\item{markers}{marker list to plot on visium HD data}
}
\value{
Seurat object with all data on markers needed for downstream analysis
}
\description{
This function pulles the read count from the visium dataset for each spot for selected markers
}
