% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindUniqueMarkers.R
\name{FindUniqueMarkers}
\alias{FindUniqueMarkers}
\title{Find Unique Markers - U method}
\usage{
FindUniqueMarkers(
  obj,
  group_by,
  p.threshold = 1,
  threshold = 0,
  P_in.thersh = 0,
  P_out.thersh = 1,
  varfeatures = NULL,
  smallcluster = NULL,
  method = "BH",
  progresstext = T,
  jumpFix = 200
)
}
\arguments{
\item{obj}{Seurat object after any clustering}

\item{group_by}{The cluster labels in metafile of obj}

\item{p.threshold}{p.value threshold for the Umethod}

\item{threshold}{what is the expression threshold that a cell is counted as positive for this gene, default is 0.}

\item{P_in.thersh}{probability of expressing inside the cluster threshold}

\item{varfeatures}{Which features to enter the analysis. For correct analysis = row.names(seurat.obj),if NULL this method will run on the 2000 most variable genes}

\item{smallcluster}{a vector of cluster names so they will be omited from the analysis, if there are small mixed clusters it is best if you ommit them here}
}
\value{
A data.frame object with the most uniquely expressed genes per cluster,that passed filtering parameters.
}
\description{
This function identifies the most unique markers per cluster in a Seurat object.
If the cluster are labeled as charachters, and not numbers, matchnames parameter shoud be F.
}
