% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindUniqueMarkers.R
\name{FindUniqueMarkers}
\alias{FindUniqueMarkers}
\title{Find Unique Markers - U Method}
\usage{
FindUniqueMarkers(
  obj,
  group_by,
  p.threshold = 1,
  threshold = 0,
  P_in.thersh = 0,
  P_out.thersh = 1,
  varfeatures = NULL,
  smallcluster = NULL,
  method = "BH"
)
}
\arguments{
\item{obj}{A Seurat object after clustering.}

\item{group_by}{The cluster labels in the metadata of \code{obj}.}

\item{p.threshold}{The p-value threshold for the U method. Defaults to a value of 0.05.}

\item{threshold}{The expression threshold above which a cell is considered positive for a gene. Default is 0.}

\item{varfeatures}{The features (genes) to use in the analysis. By default, the 2000 most variable genes will be used unless specified otherwise. If NULL, the method uses the 2000 most variable genes.}

\item{smallcluster}{A vector of cluster names to exclude from the analysis. This is useful for omitting small mixed clusters.}

\item{method}{The p-value adjustment method. Defaults to "BH" (Benjamini-Hochberg). Set to "none" for raw p-values or choose other methods available in the \code{p.adjust} function.}

\item{P_in_thresh}{The minimum probability of expressing a gene inside the cluster for filtering the final results. Default keeps the full gene list.}

\item{P_out_thresh}{The maximum probability of expressing a gene in any other cluster for filtering the final results. Default keeps the full gene list.}
}
\value{
A \code{data.frame} containing the most uniquely expressed genes per cluster that passed the filtering criteria.
}
\description{
This function identifies the most unique markers for each cluster in a Seurat object.
It supports both numerical and character-based cluster labels. If the clusters are labeled as characters, set the \code{matchnames} parameter to \code{FALSE}.
}
